Sequence names are now metadata and used in blurb
For large files this will be more useful than for small files, I imagine. Whenever a user uploads a
multi-record genbank file, they now see

        2 sequences [HE614281,HE614282]
        format:genbank, database:?

This extra information can be helpful if you're working with a collection of genbank records as a
single file. For sequences with many, many records, we've cut off the sequence listing at 10 values:

        14 sequences [HE614281,
        HE614282, HE614281,
        HE614282, HE614281,
        HE614282, HE614281,
        HE614282, HE614281,
        HE614282, ...]

Though that number could probably stand to be decreased