Ribotaper update (#361)
* add ribotaper

* add newline at end of file

* initialize create_annotation_files.xml

* initialize ribotaper.xml

* initialize create_metaplots.xml

* galaxy ribotaper: add script folder

* Ribotaper.sh: run R files with Rscript in commandline

* create_annotation_files.xml: use a file to record annotation_path

* ribotaper.xml: add input param for annotation_path

* create_annotation_files.xml: use tar ball to output dir

* create_metaplots.xml: modify command, input, output
the major change is to use gs to merge pdfs from create_metaplots.bash, then output the merged pdf.

* ribotaper.xml: modify command, input and output.
The major change is to use tarball as one input to pass annotation directory. For the outputs, the number is increased from 2 to 8 files.

* CCDS_orf_finder.R: prepend 'Rscript' so that include_multi_nomerge.R does not only rely on /usr/bin/Rscript

* NONCCDS_orf_finder.R: prepend 'Rscript'
The reason is to avoid include_multi_nomerge.R only relying on /usr/bin/Rscript

* Ribotaper_ORF_find.sh: prepend 'Rscript'
The reason is to avoid R files only relying on /usr/bin/Rscript

* create_annotations_files.bash: prepend 'Rscript'
The reason is to avoid R files only relying on /usr/bin/Rscript

* create_annotations_files.bash: prepend 'Rscript and add '/''
The reason is to avoid R files only relying on /usr/bin/Rscript. Some paths in the R file are missing '/'

* create_metaplots.xml: modify inputs for the command, find.
The previous one failed with *.pdf. It now succeeds with *.pdf.

* ribotaper.xml: add temporary tags in the help sections to guide the edition by Lorenzo.

* create_annotation_files.xml: define test output

* create_annotation_files.xml: provide help message.

* create_metaplots.xml: define test parameters and output

* ribotaper.xml: modify outputs and tests
In outputs, provide correct file names for ORFs_max and ORFs_max_filt. In tests, provide an additional parameter for .

* create_metaplots.bash: provide tag to pdf files while moving them to metaplots
The reason is to restrict file movement to the pdf files only generated by metagene analysis.

* add additional files to test-data folder for automatic test of the tools

* rename galaxy tools of ribotaper

* ribotaper conda package
built recipes for  r-multitaper,  r-xnomial, ribotaper

* galaxy ribotaper: latest status

* add the version of ribotaper

* add addtional output for the table

* ribotaper_part1_create_annotation_files.xml: update the help section.

* ribotaper_part2_create_metaplots.xml: updated the help section

* ribotaper_part3_main.xml: update the help section

* delete the obsolete files

* add metagene to test-data