Enable stdout SAM file to be written as history item
Updated <outputs> and <command> sections. When --writeMappings is set to Yes/True, a SAM-formatted file is written to stdout and redirected to its own history item (name=output_sam).

NOTE: This filter condition (line 342) seems to always evaluate to True, even when writeMappings is False. Needs to be changed (I'm not sure what to) so that a SAM file isn't produced as output if writeMappings is set to 'No'/False.