Add PhyloPhlAn tool
PhyloPhlAn is an integrated pipeline for large-scale phylogenetic profiling of genomes and metagenomes.
PhyloPhlAn GitHub: https://github.com/biobakery/phylophlan

The tool suite consists of multiple scripts which need to be wrapped into tools and data managers, but this branch
is only concerned with creating the tool wrappers. As of this commit only phylophlan.xml is largely complete.
This is the main script to run the Concatenation and Gene-trees pipelines, and allows the user to configure
which external tools they want to use at every analysis step.

- Support for some preconfigured external tools is missing (Opal, UPP and astrid).
- Using aligned markers from StrainPhlAn (--strainphlan) is not supported.
- Code and .loc files to access cached datasets are missing.

phylophlan_assign_sgbs and phylophlan_draw_metagenomic are used to report and visualize the closest species-level
genome bins, for each bin from a metagenomic assembly analysis. My progress on them has recently stalled, because the scripts
rely entirely on the presence of a cached database, which I did not implement like above.
Besides that they are missing an expanded help section, and the current release of the assign_sgbs script
has a bug limiting some of the functionality (pairwise mash distances of the input).

phylophlan_strain_finder would be a tool to perform analysis on trees and mutation rate tables build with phylophlan,
but I have not wrapped it (yet?).

The test data have been created by cutting down example data from the tutorials on the PhyloPhlAn github.