Add/Change epsilon parameter, 0.01 to 0.1
From what I understand, the metrics take into account the computation time, which is never exactly the same, so this already causes differences in the files. However, there is also a difference related to the machine that executes the computation. I tried running it locally and via Codespaces, and the difference is much larger than just running it twice on the same machine.